refactor: introduce telegram command router
